Move getTransformFileFromArgs to utils
