+++ src/utils/index.ts
@@ -1,2 +1,3 @@
 export * from "./getHelpParagraph";
+export * from "./getTransformFileFromArgs";
 export * from "./getTransforms";
